% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nv_atlas.R
\name{download_nv_atlas}
\alias{download_nv_atlas}
\title{Download Neurovault Atlas and Label file}
\usage{
download_nv_atlas(
  id,
  verbose = TRUE,
  outdir = tempfile(),
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{id}{id of the atlas}

\item{verbose}{print diagnostic messages}

\item{outdir}{output directory for images}

\item{overwrite}{Will only overwrite existing file if \code{TRUE}.}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
A \code{data.frame} of the image information
and their output filenames
}
\description{
Download Neurovault Atlas and Label file
}
\examples{
res = nv_atlas(id = 1408, verbose = TRUE)
df = results_to_df(res$content$results)
res = download_nv_atlas(id = 1408, verbose = 2)

}
