% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nv_atlas.R
\name{nv_atlas}
\alias{nv_atlas}
\alias{nv_atlas_names}
\alias{nv_atlas_df}
\title{Neurovault Atlases}
\usage{
nv_atlas(id = NULL, verbose = TRUE, secure = TRUE, ...)

nv_atlas_names(...)

nv_atlas_df(verbose = TRUE, ...)
}
\arguments{
\item{id}{id of the atlas}

\item{verbose}{print diagnostic messages}

\item{secure}{passed to \code{\link{nv_base_url}} for https}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
List of the result of the \code{\link{GET}} call and
the content
}
\description{
Returns a  list of dictionaries with information corresponding
to each atlas stored in NeuroVault.
}
\examples{
res = nv_atlas()
df = results_to_df(res$content$results)
nv_atlas_names()
nv_atlas_df()
}
