% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_base_url.R
\name{nv_base_url}
\alias{nv_base_url}
\title{Neurovault Base URL}
\usage{
nv_base_url(secure = TRUE)
}
\arguments{
\item{secure}{Should https be used (may be necessary)}
}
\value{
Character vector (length 1) of URL
}
\description{
Neurovault Base URL
}
\examples{
nv_base_url()
}
