% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_collection_id.R
\name{nv_collection_id}
\alias{nv_collection_id}
\title{Retrieve Neurovault Collections ID}
\usage{
nv_collection_id(doi = NULL, owner = NULL, name = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{doi}{Digital Object Identifier (DOI) for the project or collection}

\item{owner}{owner of the collection}

\item{name}{name of the collection}

\item{verbose}{print diagnostic messages}

\item{...}{additional options to pass to \code{\link{nv_collection}}}
}
\value{
An ID or \code{NULL} if not found
}
\description{
Retrieve Neurovault Collections ID
}
\examples{
id = nv_collection_id(name = "FeatureX IAPS Test")
\dontrun{
id = nv_collection_id(doi = "10.1016/BAD_ID")
}
}
