% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_collection_images.R
\name{nv_collection_images}
\alias{nv_collection_images}
\title{Neurovault Collections Images}
\usage{
nv_collection_images(id, verbose = TRUE, secure = TRUE, ...)
}
\arguments{
\item{id}{id of the collection}

\item{verbose}{print diagnostic messages}

\item{secure}{passed to \code{\link{nv_base_url}} for https}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
List of the result of the \code{\link{GET}} call and
the content
}
\description{
Neurovault Collections Images
}
\note{
See \url{https://neurovault.org/api-docs}
}
\examples{
res = nv_collection_images(id = 77)
df = results_to_df(res$content$results)

}
