% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_nidm_results.R
\name{nv_nidm_results}
\alias{nv_nidm_results}
\title{NIDM Results}
\usage{
nv_nidm_results(verbose = TRUE, secure = TRUE, max_count = Inf, ...)
}
\arguments{
\item{verbose}{print diagnostic messages}

\item{secure}{passed to \code{\link{nv_base_url}} for https}

\item{max_count}{Maximum count of records to call, the number of records
may be larger than this based on how the limits are set for API calls}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
List of the result of the \code{\link{GET}} call and
the content
}
\description{
Returns a list of dictionaries with information
corresponding to each nidm_results stored in NeuroVault.
}
\examples{
nidm = nv_nidm_results(max_count = 100)
# df = results_to_df(nidm$content$results)
\dontrun{
nidm = nv_nidm_results()
}
}
