% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_to_df.R
\name{results_to_df}
\alias{results_to_df}
\title{Convert Neurovault results output to a \code{data.frame}}
\usage{
results_to_df(results)
}
\arguments{
\item{results}{Results list from the content of a \code{nv} command}
}
\value{
A \code{data.frame} of the output
}
\description{
Convert Neurovault results output to a \code{data.frame}
}
\examples{
res = nv_collection(doi = "10.1016/j.neurobiolaging.2012.11.002")
results = res$content$results
df = results_to_df(results)
}
