% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nv_collection_images.R
\name{download_nv_collection_images}
\alias{download_nv_collection_images}
\title{Download Neurovault Collections Images}
\usage{
download_nv_collection_images(
  id,
  verbose = TRUE,
  outdir = tempfile(),
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{id}{id of the collection}

\item{verbose}{print diagnostic messages}

\item{outdir}{output directory for images}

\item{overwrite}{Will only overwrite existing file if \code{TRUE}.}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
A \code{data.frame} of the image information
and their output filenames
}
\description{
Download Neurovault Collections Images
}
\examples{
res = nv_collection_images(id = 77)

if (.Platform$OS.type != "windows") {
  res = download_nv_collection_images(id = 77, verbose = 2)
}
}
