% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nv_all_images.R
\name{nv_all_images}
\alias{nv_all_images}
\title{Neurovault Images}
\usage{
nv_all_images(verbose = TRUE, secure = TRUE, ...)
}
\arguments{
\item{verbose}{print diagnostic messages}

\item{secure}{passed to \code{\link{nv_base_url}} for https}

\item{...}{additional options to pass to \code{\link{GET}}}
}
\value{
List of the result of the \code{\link{GET}} call and
the content
}
\description{
Neurovault Images
}
\note{
See \url{https://neurovault.org/api-docs}
}
\examples{
\dontrun{
images = nv_all_images()
}
}
